Removal of antennas and other changes
The following edits has been made:
1) Simplification of certain critical traces
2) Removal of all antennas. Antennas are large traces connected to high
   impedance nodes. I'm quite strict about what high impedance means.
   Anything larger than a few Ohm can be considered high impedance.
3) Added labels most silkscreen labels where outside the physical board
   area
4) Added a thermal area to force

TO DO:
1) Add explanation near to all switches and trimpots
2) Check DRC
3) Add some GND vias
4) Check very carefully gerbers to see if any obvious manifacturing
   issue emerges
5) Send gerbers to fabrication.
6) Consider OPA209 or OPA140 for some of the opamps in the board.
One of the design choices of this board is that some opamps don't have
resistors for input current compensation. This is related to the choice
of the OPA189 for all board opamps (OPA189 has pA input current). It
turns out that for precision opamps manufacturers are starting to
discourage the use of resistors. Look at this [note](https://www.analog.com/en/resources/analog-dialogue/studentzone/studentzone-june-2017.html)